Fixed several bugs in logging chest access, so it works now.
1) In BBPlayerInfo, when saving the contents of just-opened chest,
the contents are probably passed just as a reference. These contents are then
modified by bukkit somewhere underneath, so when closing the chest, they
are changed to current chest contents. That means, no content changes could
be registered. We need to explicitly clone co contents, so they stay as they
are all the time. Sadly, this was doesn't clone the Data field, I haven't figured
out how to correctly retrieve it, so pleas fix it :)

2) There were some minor bugs in DeltaChest, regarding plus/minus signs, ID of
completely removed items, or items added to empty slots. It's fixed. Also,
I changed it, so it logs only the slots, that are changed. Hope it's not a problem.

3) In BBPlayerListener, I removed two calls to closeChestIfOpen, in
PlayerDropItem and PlayerPickupItem events. These events can happen even
when the player is browsing the chest right now, so it would screw logging.

4) Finally, bug that chased me all day, the chest.getInventory().getContents()
method only returns inventory of the block, that was right-clicked, that means
only half of contents of double chest. So i wrote double-chest-detection and
new method, to return chest contents.